+++ examples/multi-server/client.js
@@ -9,6 +9,7 @@ let ServiceBroker = require("../../src/service-broker");
 
 // Create broker
 let broker = new ServiceBroker({
+	namespace: "multi",
 	nodeID: process.argv[2] || "client-" + process.pid,
 	transporter: "NATS",