+++ examples/multi-server/server.js
@@ -7,6 +7,7 @@ let { MoleculerError } = require("../../src/errors");
 
 // Create broker
 let broker = new ServiceBroker({
+	namespace: "multi",
 	nodeID: process.argv[2] || "server-" + process.pid,
 	transporter: "NATS",
 	logger: console